test(alerts): split prewarm coverage from state module
